backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic-settings==2.5.2
google-genai==0.7.1
langchain-google-genai==0.2.2
chromadb==0.5.11
pypdf==5.1.0
unstructured==0.16.4
faster-whisper==1.0.3